feat: Added Event for Token

+++ appoint-ease/src/LoginForm.js
@@ -1,6 +1,12 @@
 import React, { useState } from 'react';
+import { jwtDecode } from "jwt-decode";
+import EmailConfirmationMessage from './EmailConfirmationMessage'; 
 
 const LoginForm = ({handleLogin}) => {
+
+  const [showEmailConfirmationMessage, setShowEmailConfirmationMessage] = useState(false);
+  const [loginFailed, setLoginFailed] = useState(false);
+
   const [formData, setFormData] = useState({
     username: '',
     password: '',
@@ -28,16 +34,37 @@ const LoginForm = ({handleLogin}) => {
 
       if (response.ok) {
         const jsonResponse = await response.json();
-        const { token, role, userId } = jsonResponse;
+        const token = jsonResponse.token;
+        const decodedToken = jwtDecode(token);
+        const userId = decodedToken["http://schemas.xmlsoap.org/ws/2005/05/identity/claims/nameidentifier"];
+        const username = decodedToken["http://schemas.xmlsoap.org/ws/2005/05/identity/claims/name"];
+        const role = decodedToken["http://schemas.microsoft.com/ws/2008/06/identity/claims/role"];
+        const emailConfirmation = decodedToken.EmailConfirmation;
+   
+        if(emailConfirmation === "False"){
+          setShowEmailConfirmationMessage(true); 
+           console.log("False");
 
-        localStorage.setItem('token', token);
-        localStorage.setItem('userId', userId);
-        localStorage.setItem('role', role);
-        
-        handleLogin(role, token, userId);
-        console.log('Login successful');
+        }else{
+          setShowEmailConfirmationMessage(false); 
+
+          localStorage.setItem('token', decodedToken);
+          localStorage.setItem('userId', userId);
+          localStorage.setItem('role', role);
+          localStorage.setItem('username', username);
+    
+            
+            handleLogin(role, decodedToken, userId);
+            console.log('Login successful');
+            console.log(role);
+            console.log(userId);
+        }
+
+     
+
       } else {
-        // Handle unsuccessful login, display error message, etc.
+        setShowEmailConfirmationMessage(false); 
+        setLoginFailed(true);
         console.error('Login failed');
       }
     } catch (error) {
@@ -47,9 +74,14 @@ const LoginForm = ({handleLogin}) => {
 
   return (
     <div className="container">
-      <div className="card mt-4">
-        <div className="card-body">
-          <h2 className="mb-4 text-center">Login</h2>
+      <div className="card mt-4 bg-light p-4 rounded">
+        <div className="card-body ">
+        {showEmailConfirmationMessage && (
+          <div className="alert alert-warning" role="alert">
+              <EmailConfirmationMessage/>
+          </div>
+        )}
+       <h2 className="mb-4 text-center">Login</h2>
           <form onSubmit={handleSubmit}>
             <label htmlFor="username" className="form-label">Username</label>
             <input type="text" className="form-input" id="username" name="username" value={formData.username} onChange={handleInputChange} required />
@@ -58,9 +90,13 @@ const LoginForm = ({handleLogin}) => {
             <input type="password" className="form-input" id="password" name="password" value={formData.password} onChange={handleInputChange} required />
 
             <div className="d-flex mt-3">
-              <button type="submit" className="btn btn-submit">Login</button>
+            <button type="submit" className="btn btn-primary w-100">
+              Login
+            </button>
             </div>
           </form>
+          <br></br>
+          {loginFailed && <div className="alert alert-danger mt-3" role="alert">Login failed. Please check your credentials and try again.</div>}
         </div>
       </div>
     </div>
